requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
pydantic==2.5.1
pandas==2.1.3
beautifulsoup4==4.12.2
pytest==7.4.3
alembic==1.12.1
jinja2==3.1.2
openai